Implement SSR for Form (#3407)
* Implement SSR for LayoutManager

* Implement SSR for Form

* Fix test module

* refactor

* fix tabbed item tests

* fix errors on SSR
