+++ testing/tests/DevExpress.serverSide/form.tests.js
@@ -1,0 +1,3 @@
+"use strict";
+
+require("../DevExpress.ui.widgets.form/form.markup.tests.js");
